+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/) and this
 project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.2.4] - Unreleased
+### Add
+- Tabler icons.
+
+### Fixed
+- Updated icons: `lucide`, `simple-icons`, `material-symbols`.
+
 ## [0.2.3] - 2024-08-12
 ### Fixed
 - Options type.
@@ -41,6 +48,7 @@ project adheres to [Semantic Versioning](https://semver.org/).
 ## [0.1.0] - 2024-06-15
 First version
 
+[0.2.4]: https://github.com/lumeland/icon-plugins/compare/v0.2.3...HEAD
 [0.2.3]: https://github.com/lumeland/icon-plugins/compare/v0.2.2...v0.2.3
 [0.2.2]: https://github.com/lumeland/icon-plugins/compare/v0.2.1...v0.2.2
 [0.2.1]: https://github.com/lumeland/icon-plugins/compare/v0.2.0...v0.2.1